pyproject.toml
[project]
name = "manimator"
version = "0.1.0"
description = "Create Beautiful Mathematical Animation just by prompting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio-mqtt>=0.11.0",
    "fastmcp>=0.2.0",
    "google-genai>=1.21.1",
    "google-generativeai>=0.3.0",
    "manim>=0.17.0",
    "mcp>=0.4.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
